Fixed issue #2299: Better support the git staging process
This adds the ability to stage or unstage hunks, lines or whole files from the Commit window. With the checkbox "Staging support" disabled, it all should behave exactly as it used to, apart from a minor bug in the handling of renamed files by the View Patch window, which has been fixed.

Signed-off-by: Matheus Michels <63876508+tchesftw@users.noreply.github.com>
Signed-off-by: Sven Strickroth <email@cs-ware.de>